Added the ability to delete fields with the del. Also changed the implementation of Table from using vectors to using pointers. There are just way too many strange errors that it's just easier to use a pointer array to store entries. Fixed a lot of Table bugs that I have not seen previously (it's actually embarrassing how many there are, just shows that this version still needs LOTS of testing).
Also fixed a GC bug that sweeps global variable names before they can be assigned to the globalNames table. Previously, the globalValue array was written into before the variable name, this might cause the globalValue to expand (aka run out of capacity) which might trigger the GC (or always trigger the GC if GC stress test mode is on). Since the name is not yet assigned, the GC will see it as unreachable and sweep the name. By changing the writing order this is now fixed.

P.S. Hopefully this is the final time I see a GC bug, these bugs are nightmarishly hard to find. However, I doubt it, definitely need to find some time to just test this interpreter.